Рефакторинг, добавлен метод получения среднего рейтинга, добавлен клиент сервиса товаров

+++ service/src/main/java/com/example/reactiveratingservice/entity/RatingEntity.java
@@ -9,8 +9,12 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Column;
 import org.springframework.data.relational.core.mapping.Table;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
+/**
+ * Сущность рейтинга
+ */
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
@@ -36,6 +40,5 @@ public class RatingEntity {
      * Значение рейтинга
      */
     @Column("value")
-    private Double value;
-
+    private BigDecimal value;
 }
